unused-requirements.txt
cffi
tzlocal
odfpy
defusedxml
six
diff-match-patch
python-nvd3
et-xmlfile
text-unidecode
idna
python-bidi
django-bower


asgiref
charset-normalizer
pytz
tinycss2
typing_extensions
pycparser
xlrd
tzdata
pypng
webencodings
arabic-reshaper
datetime-truncate
dj-database-url
cssselect2
cryptography
django-colorfield
django-admin-interface
requests
djangorestframework
django-import-export
python-slugify
sqlparse
django-multiselectfield
lxml
svglib
qrcode
oscrypto
chardet
asn1crypto
urllib3
colorama
pypdf
xlwt
certifi
reportlab
django-crispy-forms
django-memoize
xhtml2pdf
crispy-bootstrap5
pyhanko-certvalidator
django-compat
tablib
django-nvd3
html5lib
openpyxl
uritools
